requirements.txt
fastapi~=0.103.2
uvicorn==0.30.1
peewee==3.18.3
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
pydantic~=2.5.3
uu~=0.21.1